Refactor scrapper endpoints
Moving the scrapper endpoints to a dedicated file for better organization and maintainability.
