+++ app/api/routes/news.py
@@ -1,10 +1,8 @@
-"""Scraper for Hacker News articles."""
+"""Scraper classes for Hacker News articles."""
 
 import json
-from typing import Annotated
 
 from bindl.logger import setup_logger
-from fastapi import APIRouter, Path, status
 from playwright.async_api import (
     Browser,
     ElementHandle,
@@ -17,10 +15,6 @@ from pydantic import BaseModel, Field
 from app.dependencies.redis import get_redis_repo
 
 LOG = setup_logger(__name__)
-
-router = APIRouter()
-
-HACKER_NEWS_URL = "https://news.ycombinator.com/"
 
 
 class ScraperExceptionError(Exception):
@@ -291,49 +285,3 @@ class HackerNewsScraper(Scraper):
             response.dict(),
         )
         return response
-
-
-@router.get(
-    "/",
-    summary="Scrape Hacker News",
-    description="Scrape Hacker News articles.",
-    status_code=status.HTTP_200_OK,
-    tags=["Scraping"],
-)
-async def scrape() -> list[ScraperResponse]:
-    """Scrape Hacker News articles.
-
-    **Returns:**
-    - A list of dictionaries containing the scraped articles.
-    e.g.
-        - `/` will scrape the first page of Hacker News articles.
-    """
-    scraper = HackerNewsScraper(HACKER_NEWS_URL)
-    return await scraper.scrape()
-
-
-@router.get(
-    "/{num_pages}",
-    summary="Scrape multiple pages of Hacker News",
-    description="Scrape multiple pages of Hacker News articles. The number of pages to scrape is specified in the URL.",
-    status_code=status.HTTP_200_OK,
-    tags=["Scraping"],
-)
-async def scrape_multiple_pages(
-    num_pages: Annotated[int, Path(..., gt=0, description="Number of pages to scrape")],
-) -> list[ScraperResponse]:
-    """Scrape multiple pages of Hacker News articles.
-
-    The number of pages to scrape is specified in the URL.
-
-    **Parameters:**
-    - `num_pages`: The number of pages to scrape.
-
-    **Returns:**
-    - A list of ScraperResponse containing the scraped articles.
-    e.g.
-        - `/scrape/1` will scrape 1 page of Hacker News articles.
-        - `/scrape/2` will scrape 2 pages of Hacker News articles.
-    """
-    scraper = HackerNewsScraper(HACKER_NEWS_URL)
-    return await scraper.scrape(pages=num_pages)
